requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-health-check==3.17.0
djoser==2.1.0

